# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.DBObjectMgtCreateRequest:
    properties:
      dbmgt:
        example: 1
        type: integer
      description:
        example: Table description
        type: string
      objectName:
        example: table1
        type: string
      objectTypeId:
        example: 1
        type: integer
      sqlParam:
        example: param1=value1
        type: string
      status:
        example: active
        type: string
    type: object
  controllers.DBObjectMgtCreateResponse:
    properties:
      id:
        example: 123
        type: integer
      message:
        example: Object was created successfully
        type: string
    type: object
  controllers.DBObjectMgtDeleteResponse:
    properties:
      message:
        example: Object was deleted
        type: string
    type: object
  controllers.DBObjectMgtItem:
    properties:
      dbmgt:
        example: 1
        type: integer
      description:
        example: Table description
        type: string
      id:
        example: 1
        type: integer
      objectName:
        example: table1
        type: string
      objectTypeId:
        example: 1
        type: integer
      sqlParam:
        example: param1=value1
        type: string
      status:
        example: active
        type: string
    type: object
  controllers.DBObjectMgtListResponse:
    properties:
      message:
        items:
          $ref: '#/definitions/controllers.DBObjectMgtItem'
        type: array
    type: object
  controllers.DBObjectMgtUpdateResponse:
    properties:
      message:
        example: Object was updated
        type: string
    type: object
...
  controllers.DatabaseConnectionErrorResponse:
    properties:
      error:
        example: database connection failed
        type: string
    type: object
...
  controllers.InternalServerErrorResponse:
    properties:
      error:
        example: database connection failed
        type: string
    type: object
...
  controllers.ObjectCreationErrorResponse:
    properties:
      error:
        example: failed to create object
        type: string
    type: object
...
  controllers.StandardErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
info:
  contact: {}
  description: DBF Artifact API
  title: dbfartifactapi
  version: "1.0"
paths:
...
  /api/queries/dbobjectmgt:
    post:
      consumes:
      - application/json
      description: Creates a new database object management entry with specified parameters
      parameters:
      - description: Database Object Management object
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controllers.DBObjectMgtCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Object created successfully
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtCreateResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ObjectCreationErrorResponse'
      summary: Create database object management
      tags:
      - DB Object Management
  /api/queries/dbobjectmgt/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing database object management entry by ID
      parameters:
      - description: Object Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Object deleted successfully
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtDeleteResponse'
        "400":
          description: Invalid object management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Object not found
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Delete database object management
      tags:
      - DB Object Management
    put:
      consumes:
      - application/json
      description: Updates an existing database object management entry by ID
      parameters:
      - description: Object Management ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Database Object Management object
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controllers.DBObjectMgtCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Object updated successfully
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtUpdateResponse'
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Object not found
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Update database object management
      tags:
      - DB Object Management
  /api/queries/dbobjectmgt/cntmgt/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all database objects associated with the specified connection
        management ID
      parameters:
      - description: Connection Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of database objects
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtListResponse'
        "400":
          description: Invalid connection management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.DatabaseConnectionErrorResponse'
      summary: Get database objects by connection management ID
      tags:
      - DB Object Management
  /api/queries/dbobjectmgt/dbmgt/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all database objects associated with the specified database
        management ID
      parameters:
      - description: Database Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of database objects
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtListResponse'
        "400":
          description: Invalid database management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.DatabaseConnectionErrorResponse'
      summary: Get database objects by database management ID
      tags:
      - DB Object Management
...
swagger: "2.0"